+++ FootballData/Models/TeamResult.cs
@@ -4,11 +4,15 @@ namespace FootballData.Models;
 
 public class TeamResult
 {
-    [JsonPropertyName("country")] public string Country { get; set; }
+    [JsonPropertyName("country")] 
+    public string Country { get; set; }
 
-    [JsonPropertyName("code")] public string Code { get; set; }
+    [JsonPropertyName("code")] 
+    public string Code { get; set; }
 
-    [JsonPropertyName("goals")] public long Goals { get; set; }
+    [JsonPropertyName("goals")] 
+    public int Goals { get; set; }
 
-    [JsonPropertyName("penalties")] public long Penalties { get; set; }
+    [JsonPropertyName("penalties")] 
+    public int Penalties { get; set; }
 }
